requirements.txt
git+https://github.com/pyrogram/pyrogram
TgCrypto
git+https://github.com/pytgcalls/pytgcalls
python-dotenv
youtube_dl
youtube_search_python
OpenCC
Pillow
requests
youtube_search
aiohttp
aiofiles
wget
python-arq
